@@ -1,14 +1,15 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.0.3'
-gem 'active_model_serializers'
+gem 'rails', '4.2.0'
+gem 'active_model_serializers', '0.9.3'
 gem 'sass-rails', '~> 4.0.0'
 gem 'bootstrap-sass', '~> 3.2.0'
 gem 'haml'
 gem 'haml-rails'
 gem 'coffee-rails', '~> 4.0.0'
 gem 'jquery-rails'
+gem 'responders', '~> 2.0'
 
 group :production do
   # Use postgresql as the database for Active Record
