@@ -1,7 +1,4 @@
 class UserSerializer < ActiveModel::Serializer
-  attributes :username, :password
-  def username
-    object.username
-  end
+  attributes :id, :username
 
 end
