@@ -1,5 +1,17 @@
 Todo::Application.routes.draw do
-  resources :users do 
+
+  namespace :api, defaults: {format: :json} do
+    resources :users do
+      resources :lists
+    end
+    
+    resources :lists, only: [] do
+      resources :items, only: [:create]
+    end
+    resources :items, only: [:destroy]
+  end
+
+  resources :users do
     resources :lists, except: [:index]
   end
 
